Gateway Fixed

+++ Recommend-Service/src/main/java/com/ust/Recommend_Service/Model/Recommend.java
@@ -13,5 +13,7 @@ import lombok.NoArgsConstructor;
 public class Recommend {
     @Id
     private String movieId;
-    private float rating;
+    private Double rating;
+
+
 }
